- Added possibility for consumers to *not* affect the presentation time (adjustments in output.cpp made based on the buffer_depths of the consumers) of other consumers. This is good when consumers that are usually spontaneously added and removed during playout (image_consumer, ffmpeg_consumer, streaming_consumer and channel_consumer) are used, because it prevents the lag previously experienced when adding/removing such consumers. - Fixed streaming_consumer sometimes crashing the server during destruction.

+++ core/consumer/frame_consumer.cpp
@@ -132,7 +132,7 @@ public:
 		return consumer_->has_synchronization_clock();
 	}
 
-	virtual size_t buffer_depth() const override
+	virtual int buffer_depth() const override
 	{
 		return consumer_->buffer_depth();
 	}
@@ -157,7 +157,7 @@ const safe_ptr<frame_consumer>& frame_consumer::empty()
 		virtual int64_t presentation_frame_age_millis() const { return 0; }
 		virtual std::wstring print() const override {return L"empty";}
 		virtual bool has_synchronization_clock() const override {return false;}
-		virtual size_t buffer_depth() const override {return 0;};
+		virtual int buffer_depth() const override {return 0;};
 		virtual int index() const{return -1;}
 		virtual boost::property_tree::wptree info() const override
 		{
