+++ modules/decklink/consumer/decklink_consumer.cpp
@@ -587,7 +587,7 @@ public:
 		return info;
 	}
 
-	virtual size_t buffer_depth() const override
+	virtual int buffer_depth() const override
 	{
 		return config_.buffer_depth();
 	}
