+++ modules/oal/consumer/oal_consumer.cpp
@@ -154,7 +154,7 @@ public:
 		return info;
 	}
 	
-	virtual size_t buffer_depth() const override
+	virtual int buffer_depth() const override
 	{
 		return 6;
 	}
