+++ modules/ogl/consumer/ogl_consumer.cpp
@@ -581,7 +581,7 @@ public:
 		return false;
 	}
 	
-	virtual size_t buffer_depth() const override
+	virtual int buffer_depth() const override
 	{
 		return 1;
 	}
